requirements.txt
async-generator==1.10
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
cryptography==36.0.0
Flask==2.0.2
gcloud==0.17.0
gensim==4.1.2
googleapis-common-protos==1.53.0
gunicorn==20.1.0
h11==0.12.0
httplib2==0.20.1
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
jws==0.1.3
MarkupSafe==2.0.1
mecab-python3==1.0.4
numpy==1.21.3
oauth2client==3.0.0
outcome==1.1.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
pyOpenSSL==21.0.0
pyparsing==2.4.7
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.7.2
scipy==1.7.1
selenium==4.1.0
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.10.0.2
unidic-lite==1.0.8
urllib3==1.26.7
Werkzeug==2.0.2
wsproto==1.0.0
zipp==3.6.0